[TASK] Avoid temporary files in .ddev/ folder
Currently, files/folders are copied to a folder and this folder is mounted into the container. This is causing a mess in the .ddev folder. Which files to commit?

To prevent this, the ddev command is now a host command
and copies all required files directly to the container volume.

Fixes #1